CastScraper

Description

This Module performs http requests to get the HTML in return. It then parses the HTML to extract Actor Info

Documentation

type ActorName = String

type URL = String

data Actor

Constructors

Actor

Fields

name :: String

dob :: String

nationality :: String

cast_gender :: String

role :: String

urlBuilder :: ActorName -> URL

This function builds the url using the ActorName in the correct format. This helps render Actor info dynamically. It returns the url to be scraped and parsed.

getTags :: URL -> IO [Tag String]

This function uses the tagsoup module to parse the requested HTML to a list of tags. it takes in the url as a string it returns a list of tags in the form, eg. "hellomy&amp;/world" == [TagOpen "hello" [],TagText "my&",TagClose "world"]

getActorName :: ActorName -> IO String

This function traverses through the html(parsed as a list of tags) and extracts the actor's name

getActorDOB :: ActorName -> IO String

This function traverses through the html(parsed as a list of tags) and extracts the actor's date of birth

getActorNationality :: ActorName -> IO String

This function traverses through the html(parsed as a list of tags) and extracts the actor's date of nationality

getActorGender :: ActorName -> IO String

This function traverses through the html(parsed as a list of tags) and extracts the actor's date of gender

getActorRole :: ActorName -> IO String

This function traverses through the html(parsed as a list of tags) and extracts the actor's role in harry potter

getActorInfo :: ActorName -> IO Actor

This function takes all the extracted info from the parsed HTML and creates a new actor record The function takes in an actor name as argument in the format "firstName_LastName". eg. Emma_Watson it returns an Actor