CharacterScraper

Description

This Module performs http requests to get the HTML in return. It then parses the HTML to extract Character Info

Documentation

type CharacterName = String

type URL = String

data Character

Constructors

Character

Fields

fname :: String

gender :: String

bloodStatus :: String

house :: String

patronus :: String

urlBuilder :: CharacterName -> URL

This function builds the url using the CharacterName in the correct format. This helps render Actor info dynamically. It returns the url to be scraped and parsed.

getTags :: URL -> IO [Tag String]

This function uses the tagsoup module to parse the requested HTML to a list of tags. it takes in the url as a string it returns a list of tags in the form, eg. "hellomy&amp;/world" == [TagOpen "hello" [],TagText "my&",TagClose "world"]

getFullName :: CharacterName -> IO String

This function traverses through the html(parsed as a list of tags) and extracts the Character's full name

getGender :: CharacterName -> IO String

This function traverses through the html(parsed as a list of tags) and extracts the Character's date of gender

getBloodStatus :: CharacterName -> IO String

This function traverses through the html(parsed as a list of tags) and extracts the Character's blood Status, eg. "Half-Blood"

getHouse :: CharacterName -> IO String

This function traverses through the html(parsed as a list of tags) and extracts the Character's House, eg. Gryffindor

getPatronus :: CharacterName -> IO String

This function traverses through the html(parsed as a list of tags) and extracts the Character's patronus, eg. "stag"

newCharacter :: CharacterName -> IO Character

This function takes all the extracted info from the parsed HTML and creates a new Character record The function takes in an Character name as argument in the format "firstName_LastName". eg. Harry_Potter it returns an Character