ToDB

Description

This Module contains function that talks to the database. It contains functions that stores the parse data from website into an SQlite3 database using HDBC It also contains functions that queries the database to get interesting information

Documentation

initialiseCastDB :: Connection -> IO ()

This function initialises the cast table if doesn't exist The table has 6 columns with the cast_id being the primary KEY It relates to the character table with the role of the actors

initialiseCharacterDB :: Connection -> IO ()

This function initialises the character table if doesn't exist The table has 6 columns with the character_id being the primary KEY

data CharacterFromDB

CharacterFromDB is a different type than the data type Character that is imported from CharacterScraper. The Character data type from CharacterScraper doesn't have a unique ID CharacterFromDB has an id function

Constructors

CharacterDB

Fields

id :: Integer

fname :: String

gender :: String

bloodStatus :: String

house :: String

patronus :: String

data ActorFromDB

ActorFromDB is a different type than the data type Actor that is imported from CastScraper. The Actor data type from CastScraper doesn't have a unique ID ActorFromDB has an id function

Constructors

ActorDB

Fields

cast_id :: Integer

name :: String

dob :: String

nationality :: String

cast_gender :: String

role :: String

characterToSqlValues :: Character -> [SqlValue]

This function converts Character (imported from CharacterScraper) to list of Sql Values

sqlValuesToCharacter :: [SqlValue] -> CharacterFromDB

This function converts list of Sql values to CharacterFromDB data type

prepareInsertCharacterIOStmt :: Connection -> IO Statement

This function simply prepares an insert statement for the character table

batchInsertCharacter :: [Character] -> IO ()

This function inserts list of Characters parsed from the website into the database

fetchAllValuesCharacter :: Int -> IO [CharacterFromDB]

This function uses a SELECT sql statement to query all the data from the character table using a particular id

characterNotInDB :: Int -> IO Bool

This function takes a character id as an input argument and returns true if the id doesn't exist in the database

fetchAllCharactersFromHouse :: String -> IO [CharacterFromDB]

This function uses a SELECT sql statement to query all the data from the character table with a particular House

fetchAllCharactersFromBloodStatus :: String -> IO [CharacterFromDB]

This function uses a SELECT sql statement to query all the data from the character table with a particular Blood Status

actorToSqlValues :: Actor -> [SqlValue]

This function converts Actor (imported from CastScraper) to list of Sql Values

sqlValuesToActor :: [SqlValue] -> ActorFromDB

This function converts list of Sql values to ActorFromDB data type

prepareInsertActorIOStmt :: Connection -> IO Statement

This function simply prepares an insert statement for the character table

batchInsertActor :: [Actor] -> IO ()

This function inserts list of Actors parsed from the website into the database

actorNotInDB :: Int -> IO Bool

This function takes a cast id as an input argument and returns true if the id doesn't exist in the database

fetchAllValuesActor :: Int -> IO [ActorFromDB]

This function uses a SELECT sql statement to query all the data from the actor table using a particular id