UserMenu

Description

This module basically creates the user interface that calls and utilizes all the functions imported from CharacterScraper, CastScraper and ToDB modules It also has some useful functions that beautifies the user interface on the console

Documentation

mainActors :: [[Char]]

downloadCastInfo :: IO ()

This function first calls the getActorInfo and newCharacter functions from CastScraper and CharacterScraper modules respectively, to download information from the website It then calls the batchInsertActor and batchInsertCharacter functions from ToDB module to create and populate the sqlite database.

deleteRecordFromDatabase :: IO ()

This function performs a DELETE sql statement to clear all the data from the two table

printActor :: ActorFromDB -> IO ()

This functions prints an Actor information queried from the database in a beautiful and structured way.

printCharacter :: CharacterFromDB -> IO ()

This functions prints a Character information queried from the database in a beautiful and structured way.

searchByActorName :: IO ()

This functions queries the cast table to fetch info about an actor with a particular chosen by the user

searchByCharacterName :: IO ()

This functions queries the character table to fetch info about a character with a particular chosen by the user

searchByCharacterHouse :: IO [()]

This functions queries the character table to fetch info about all the characters with a particular house chosen by the user

searchByCharacterBloodStatus :: IO [()]

This functions queries the character table to fetch info about all the characters with a particular blood status chosen by the user

userMenu :: IO ()

This function provides the user with options to query the database in a particular way. It's an interactive user interface that only terminates if the user press 0 and Enter